% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classic_ci.R
\name{classic_ci}
\alias{classic_ci}
\title{Classic Confidence Intervals for Binomial Estimate}
\usage{
classic_ci(x, n, alpha = 0.05)
}
\arguments{
\item{x}{, number of successes}

\item{n}{, number of trials}

\item{alpha}{, alpha}
}
\value{
Confidence interval as a list of lowerbound, upperbound
}
\description{
Calculates the alpha% confidence interval for a binomial estimate.
}
\examples{
x=5; n=15; classic_ci(x,n)

}
