% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffConf.R
\name{diffConf}
\alias{diffConf}
\title{Confidence Interval for the True Difference Between Two Proportions}
\usage{
diffConf(tab, a, b, FWER = 0.05, ntest = (length(tab) * (length(tab) - 1))/2)
}
\arguments{
\item{tab}{, a one way table}

\item{a}{, index for the fist estimate}

\item{b}{, index for the second estimate (greater than a)}

\item{FWER}{, family wise error rate}

\item{ntest}{, number of multiple comparisons (default value assumes all pairwise comparisons checked)}
}
\value{
List of estimate of the difference, lower bound of confidence interval, and upperbound of confidence interval
}
\description{
Takes in a one way table and estimates the difference between the true proportions of 2 categories. Uses Bonferroni correction.
}
\examples{
tab=c("x"=1, "y"=2, "z"=3); a=1; b=3; diffConf(tab, a, b)

}
