% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect2xR.R
\name{expect2xR}
\alias{expect2xR}
\title{Expected Values of a Two-Way Table}
\usage{
expect2xR(df)
}
\arguments{
\item{df}{, a data frame with  2 rows and R columns}
}
\value{
A dataframe
}
\description{
Calculates the expected values for a multinomial 2xR data frame
}
\examples{


}
