% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBoot.R
\name{myBoot}
\alias{myBoot}
\title{Bootstrap From Scratch}
\usage{
myBoot(data, iter, alpha = 0.2)
}
\arguments{
\item{data}{, a data frame where the first column has dependent variable and all other columns are independent}

\item{iter}{, number of bootstrap iterations}

\item{alpha}{, alpha level}
}
\value{
, histograms for each beta estimate, confidence intervals for each beta estimate, summary statistics 
the iter length list of beta estimates
}
\description{
Bootstrap From Scratch
}
\examples{
clerical<-data.frame(Hours=c(128.5, 113.6, 146.6, 124.3, 100.4, 119.2, 109.5, 128.5, 131.2, 112.2),
Mail=c(7781,7004,7267,2129,4878,3999,11777,5764,7392,8100),
Gifts=c(100,110,61,102,45,144,123,78,172,126),
Charge=c(886,962,1342,1153,803,1127,627,748,876,685),
Returns=c(235,388,398,457,577,345,326,161,219,287),
Checks=c(644,589,1081,891,537,563,402,495,823,555)
)

myBoot(clerical, 1000)

}
