% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_confidenceInts.R
\name{print_confidenceInts}
\alias{print_confidenceInts}
\title{Print Confidence Intervals}
\usage{
print_confidenceInts(ci)
}
\arguments{
\item{ci}{, a list or vector of length 2}
}
\value{
A string
}
\description{
Takes in a list of 2 numbers and prints it as (a,b)
}
\examples{
print_confidenceInts(c(2,3))

}
