% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printConfidenceInts.R
\name{printConfidenceInts}
\alias{printConfidenceInts}
\title{Print Confidence Intervals}
\usage{
printConfidenceInts(ci)
}
\arguments{
\item{ci}{, a list or vector of length 2}
}
\value{
A string
}
\description{
Takes in a list of 2 numbers and prints it as (a,b)
}
\examples{
printConfidenceInts(c(2,3))

}
