% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyKnots.R
\name{shinyKnots}
\alias{shinyKnots}
\title{Open the app that displays piecewise regression with 2 knots for spruce data}
\usage{
shinyKnots()
}
\value{

}
\description{
Open the app that displays piecewise regression with 2 knots for spruce data
}
\examples{
\dontrun{ shinyKnots()}

}
