% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilsonCI.R
\name{wilsonCI}
\alias{wilsonCI}
\title{Wilson Confidence Intervals for Binomial Estimate}
\usage{
wilsonCI(x, n, alpha = 0.05)
}
\arguments{
\item{x}{, number of successes}

\item{n}{, number of trials}

\item{alpha}{, alpha}
}
\value{
Confidence interval as a list of lowerbound, upperbound
}
\description{
Calculates the alpha% confidence interval for a binomial estimate using the Wilson correction.
}
\examples{
 x=5; n=15; wilsonCI(x,n)

}
